requirements.txt
apt-xapian-index==0.45
chardet==2.0.1
Cheetah==2.4.4
click==6.7
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
decorator==3.4.0
Flask==0.12.1
html5lib==0.999
imutils==0.4.2
ipython==1.2.1
itsdangerous==0.24
Jinja2==2.9.6
jsonpatch==1.3
jsonpointer==1.0
Landscape-Client==14.12
MarkupSafe==1.0
matplotlib==1.3.1
nose==1.3.1
numpy==1.12.1
oauth==1.0.1
opencv-python==3.2.0.7
PAM==0.4.2
pexpect==3.1
Pillow==2.3.0
prettytable==0.7.2
pycurl==7.19.3
pyOpenSSL==0.13
pyparsing==2.0.1
pyserial==2.6
pytesseract==0.1.6
python-apt===0.9.3.5ubuntu2
python-dateutil==1.5
python-debian===0.1.21-nmu2ubuntu2
pytz==2012rc0
PyYAML==3.10
requests==2.2.1
scikit-image==0.9.3
scipy==0.13.3
simplegeneric==0.8.1
six==1.5.2
ssh-import-id==3.21
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
urllib3==1.7.1
virtualenv==15.1.0
Werkzeug==0.12.1
zope.interface==4.0.5
